// lib/types/salesRepresentitives.ts
export interface SalesRepresentative {
    uId: string;
    name: string;
    email: string;
    password?: string;
    companyName: string;
    companyEmail: string;
    phoneNumber: string;
    country: string;
    countryKey: string;
    state: string;
    address: string;
    visaPrice: string;
    dateOfRegister: string;
  } 